Moved admin page from components to pages
