+++ django_react_starter/frontend/src/pages/AdminPage.js
@@ -2,8 +2,8 @@ import React from 'react';
 import axios from "axios";
 import { Button, Card, List, Skeleton, Stack } from '@mui/material';
 import { DataGrid, GridActionsCellItem, GridRowParams } from '@mui/x-data-grid';
-import NavBar from './NavBar';
-import SmallUserCard from './UserCards/SmallUserCard';
+import NavBar from '../components/NavBar';
+import SmallUserCard from '../components/UserCards/SmallUserCard';
 import PublishIcon from '@mui/icons-material/Publish';
 import DeleteIcon from '@mui/icons-material/Delete';
 import _ from 'lodash';
